requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
nemo_toolkit[asr]>=2.0
kokoro-onnx>=0.4
httpx>=0.27
pydantic>=2
numpy>=1.26
soundfile>=0.12
